--- cmd/meroxa/turbine/ruby/upgrade.go
@@ -1,7 +1,0 @@
-package turbinerb
-
-import "fmt"
-
-func (t *turbineRbCLI) Upgrade(_ bool) error {
-	return fmt.Errorf("command unavailable")
-}
